# perf(expr): optimize lower/upper/trim/md5