--- a/src/expr/src/vector_op/trim_characters.rs
+++ b/src/expr/src/vector_op/trim_characters.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use itertools::Itertools;
 use risingwave_common::array::{StringWriter, WrittenGuard};
 
 use crate::Result;
@@ -21,26 +20,20 @@ macro_rules! gen_trim {
     ($( { $func_name:ident, $method:ident }),*) => {
         $(#[inline(always)]
         pub fn $func_name(s: &str, characters: &str, writer: StringWriter<'_>) -> Result<WrittenGuard> {
-            let slices = characters.chars().collect_vec();
+            let pattern = |c| characters.chars().any(|ch| ch == c);
             // We remark that feeding a &str and a slice of chars into trim_left/right_matches
             // means different, one is matching with the entire string and the other one is matching
             // with any char in the slice.
-            Ok(writer.write_ref(s.$method(&slices[..])))
+            Ok(writer.write_ref(s.$method(pattern)))
         })*
     }
 }
 
-macro_rules! for_three_variants {
-    ($macro:ident) => {
-        $macro! {
-            { trim_characters, trim_matches },
-            { ltrim_characters, trim_start_matches },
-            { rtrim_characters, trim_end_matches }
-        }
-    };
+gen_trim! {
+    { trim_characters, trim_matches },
+    { ltrim_characters, trim_start_matches },
+    { rtrim_characters, trim_end_matches }
 }
-
-for_three_variants! { gen_trim }
 
 #[cfg(test)]
 mod tests {